Removed old sound files & moved audio methods into a seperate module
